refactor(llm): move openai tool_search_logging into model_constructors
